+++ workflows_definition/src/sdk/sdkconfiguration.py
@@ -19,9 +19,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.47.0'
-    gen_version: str = '2.142.2'
-    user_agent: str = 'speakeasy-sdk/python 1.47.0 2.142.2 1.0.0 openapi'
+    sdk_version: str = '1.47.1'
+    gen_version: str = '2.143.2'
+    user_agent: str = 'speakeasy-sdk/python 1.47.1 2.143.2 1.0.0 openapi'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: